requirements.txt
asgiref
automium
automium-web
blinker
certifi
charset-normalizer
click
colorama
crispy-bootstrap5
Django
django-allauth
django-ckeditor
django-cors-headers
django-crispy-forms
django-environ
django-extensions
django-js-asset
Flask
idna
itsdangerous
Jinja2
legacy
MarkupSafe
pillow
psycopg2
psycopg2-binary
requests
sqlparse
tzdata
urllib3
utils
Werkzeug
whitenoise
Gunicorn
WhiteNoise
dj-database-url
django-rosetta
scipy
nltk
pandas
django-parler
django-form-surveys
pyasynchat
python-dotenv
django-bootstrap-datepicker-plus
django-extra-views
django-db-signals
automium_web
django-otp
six
dnspython
bootstrap4